feat: Add documentation and code cleanup improvements
- Add comprehensive README.md for axum-example with usage instructions and test explanation
- Convert informal XXX comment to proper TODO with GitHub issue link
- Remove unused MockServer::with_config() method from test suite
- All changes maintain code quality and follow project coding standards
