+++ .mise.toml
@@ -9,7 +9,7 @@ cargo = { binstall = true }
 [tools]
 "cargo:cargo-audit" = "latest"
 "cargo:cargo-nextest" = "0.9"
-"cargo:git-cliff" = "2.8"
+"cargo:git-cliff" = "latest"
 "cargo:release-plz" = "latest"
 "npm:@anthropic-ai/claude-code" = "latest"
 "npm:@stoplight/spectral-cli" = "latest"
